ci: configure GitHub Actions to run Ansible checks
Configure the GitHub Actions workflow to run the Ansible syntax checks
defined in the .travis.yml file. Because we are not using Travis CI as
our build service, the configuration file can be removed from the
project.

The call to the Ansible Galaxy notification API will be implemented at a
later stage.